## Summary
Successfully moved and integrated the TrackAnalyzer component into the TestConfigurationTab section. Here's what was accomplished:

### Changes Made:

1. **Refactored TrackAnalyzer component** (`apps/web/src/components/TrackAnalyzer.tsx`):
   - Made it flexible to work both as a modal and embedded component
   - Added `asModal` prop to control rendering mode
   - Added props for external control of loading and search response
   - Improved styling consistency with Material-UI Paper components

2. **Updated TestConfigurationTab** (`apps/web/src/components/TestConfigurationTab.tsx`):
   - Removed duplicate rendering logic
   - Now uses the TrackAnalyzer component directly in embedded mode
   - Maintains all original functionality with cleaner code

3. **Updated PlexTrack component** (`apps/web/src/components/PlexTrack.tsx`):
   - Added `asModal` prop to maintain existing modal behavior
   - No breaking changes to existing functionality

### Verification:
- ✅ Type checking passes: `pnpm -r run type-check`
- ✅ Linting passes: `pnpm -r run lint` (only warnings, no errors)

The TrackAnalyzer is now properly integrated into the TestConfigurationTab section and can be referenced from there, while still maintaining backward compatibility with its modal usage in PlexTrack.
